+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 ## master
 <!-- Your comment below here -->
 
-
+* Add support for `pull_request_target` in GitHub actions. - [@chesire](https://github.com/chesire)
 
 <!-- Your comment above here -->
 ## 8.1.0