+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ## master
 <!-- Your comment below here -->
 
+* Add support for Concourse-CI [@matthewellis](https://github.com/matthewellis)
 * Add support for `pull_request_target` in GitHub actions. - [@chesire](https://github.com/chesire)
 
 <!-- Your comment above here -->