+++ CHANGELOG.md
@@ -11,10 +11,14 @@
 ## master
 <!-- Your comment below here -->
 
+
+<!-- Your comment above here -->
+
+## 8.2.0
+
 * Add support for Concourse-CI [@matthewellis](https://github.com/matthewellis)
 * Add support for `pull_request_target` in GitHub actions. - [@chesire](https://github.com/chesire)
 
-<!-- Your comment above here -->
 ## 8.1.0
 
 * correct the usage of double splat to fix Ruby 2.7 warning